# docs/api.yaml
openapi: 3.1.0
info:
  title: Parade State Management System API
  version: 0.1.0
  description: |
    API for the Battalion Parade State Management System.

    ## Authentication (issue 31)
    All endpoints (except `/auth/*`) derive the caller's identity from the
    authenticated session — the HttpOnly `session_token` cookie set by the
    Google OAuth callback, or the same token as a Bearer header for
    non-browser clients. No endpoint accepts identity from query parameters
    or request bodies; provenance (`created_by`, `updated_by`, …) is stamped
    server-side from the session user. Unauthenticated calls return 401;
    calls below the endpoint's role tier or outside the caller's scope
    grants return 403. Role is re-read from the database on every request,
    so role changes take effect immediately.

    ## Access control
    Two enforcement dimensions applied on every data endpoint:
    - **Row visibility**: user sees only personnel rows covered by their nominal-roll subunit assignments.
    - **Column visibility**: response includes only columns whose sensitivity label is met by the user's access level.
      The `columns` manifest on list endpoints is always pre-scoped — clients render headers from the manifest, never from hardcoded column names.

    Admin and super_admin roles bypass both dimensions and see all rows and columns.

    ## Column manifest pattern (Option C)
    Endpoints returning personnel data include a `columns` array alongside `rows`.
...

...
servers:
...
paths:
...
  /sessions/{sessionId}/attendance:
    parameters:
      - $ref: '#/components/parameters/SessionIdParam'
    get:
      tags: [attendance]
      summary: Get parade state — column manifest + scoped rows
      description: |
        Returns `columns` (user-visible column manifest) and `rows` (one per visible person).
        Row objects contain only keys present in the columns manifest.
        Rows are filtered to the user's subunit scope for this grouping.
        Admins receive all columns and all rows.

        Optional `subunit` filter narrows rows to a specific subunit within scope.
        Optional `status` filter narrows by one of the nine attendance statuses.

        This is the primary attendance-taking and parade state view endpoint.
      parameters:
        - name: subunit
          in: query
          description: Filter by subunit (within user scope)
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [present, absent, time_off, mc, yet_to_inpro, outpro, reporting_sick, late, att_out]
        - name: sortBy
          in: query
          schema:
            type: string
            description: Canonical or raw column name from the manifest
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceView'
        '403':
          $ref: '#/components/responses/Forbidden'
  /sessions/{sessionId}/attendance/{persNo}:
    parameters:
      - $ref: '#/components/parameters/SessionIdParam'
      - name: persNo
        in: path
        required: true
        schema:
          type: string
    patch:
      tags: [attendance]
      summary: Update attendance record (inline edit)
      description: |
        Writable fields for scoped users: `status`, `remarks`, `notesText`.
        `notesText` triggers a write-back to grouping_notes and updates
        notes_snapshot on this record.
        Unit+subunit snapshot update follows the validity-range rule:
          - If now() within grouping valid_from/valid_until: snapshots are updated.
          - If outside validity range: snapshots are NOT updated.
        Writes `attendance.updated` (and `notes.updated` if notesText changed) to audit log.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceUpdateRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceRecord'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Concurrent write conflict (last-write-wins; this is informational only)
...
  /events/attendance:
    get:
      tags: [events]
      summary: SSE stream for attendance stale detection
      description: |
        Server-Sent Events stream. The server emits a `data_changed` event when any
        attendance record within the authenticated user's current scope
        is modified by a different user.

        The client should:
        1. Open this connection on entering the attendance view.
        2. On receiving `data_changed`, show a "Refresh" prompt to the user.
        3. On user confirming refresh, reload the attendance view.
        4. Close the connection on leaving the attendance view.

        The server does NOT push full record data over this stream — only change signals.
        Connection is authenticated via session cookie. The server closes the connection
        if the session expires.

        Event format:
        ```
        event: data_changed
        data: {"updatedAt": "..."}
        ```

        Keep-alive ping every 30s:
        ```
        event: ping
        data: {}
        ```
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: SSE stream opened
          content:
            text/event-stream:
              schema:
                type: string
        '403':
          $ref: '#/components/responses/Forbidden'

  # ===========================================================================
  # Deferments (super-admin only)
  # ===========================================================================
...
components:
...
  parameters:
...
    SessionIdParam:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Forbidden:
      description: Insufficient access level or role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true

    # --- Nominal Roll ---
...
    AttendanceView:
      type: object
      description: |
        Column-manifest + rows response. Clients render table headers from `columns`
        and row cells from `rows[n][column.key]`. Keys not in the manifest are never
        present in row objects.
      properties:
        sessionId:
          type: string
          format: uuid
        groupingId:
          type: string
          format: uuid
        sessionDate:
          type: string
          format: date
        sessionType:
          type: string
          enum: [AM, PM]
        columns:
          type: array
          description: Ordered list of columns visible to this user
          items:
            $ref: '#/components/schemas/ColumnManifestEntry'
        rows:
          type: array
          items:
            $ref: '#/components/schemas/AttendanceRow'
        meta:
          type: object
          properties:
            totalRows:
              type: integer
            presentCount:
              type: integer
            absentCount:
              type: integer
    ColumnManifestEntry:
      type: object
      properties:
        key:
          type: string
          description: The key used in row objects (canonical name or raw name if unmapped)
        label:
          type: string
          description: Display label for column header
        type:
          type: string
          enum: [text, integer, date, boolean, numeric]
        isAttendanceField:
          type: boolean
          description: True for status, remarks, notes — these are inline-editable by scoped users
        displayOrder:
          type: integer
    AttendanceRow:
      type: object
      description: |
        Keys are the `key` values from the columns manifest. Always present:
        persNo, rank, fullName, unit, subUnit1/2/3, status, remarks, notes.
        All other keys depend on the user's access level.
      properties:
        persNo:
          type: string
        status:
          type: string
          enum: [present, absent, time_off, mc, yet_to_inpro, outpro, reporting_sick, late, att_out]
          default: absent
        remarks:
          type: string
        notes:
          type: string
          description: Live canonical notes (not snapshot) for display
        notesUpdatedAt:
          type: string
          format: date-time
          nullable: true
        notesUpdatedBy:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          nullable: true
      additionalProperties:
        description: Additional personnel fields visible to this user per column manifest
    AttendanceRecord:
      type: object
      description: Single attendance record as returned after a PATCH
      properties:
        id:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        persNo:
          type: string
        status:
          type: string
          enum: [present, absent, time_off, mc, yet_to_inpro, outpro, reporting_sick, late, att_out]
          default: absent
        remarks:
          type: string
        notesSnapshot:
          type: string
        unitSnapshot:
          type: string
          nullable: true
        subUnit1Snapshot:
          type: string
          nullable: true
        subUnit2Snapshot:
          type: string
          nullable: true
        subUnit3Snapshot:
          type: string
          nullable: true
        snapshotTakenAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
          nullable: true
    AttendanceUpdateRequest:
      type: object
      description: At least one of status, remarks, notesText must be present
      properties:
        status:
          type: string
          enum: [present, absent, time_off, mc, yet_to_inpro, outpro, reporting_sick, late, att_out]
        remarks:
          type: string
          maxLength: 500
        notesText:
          type: string
          maxLength: 2000
          description: |
            If present, triggers write-back to grouping_notes and updates
            notes_snapshot on this attendance record.

    # --- Audit log ---
...